requirements.txt
streamlit
streamlit-drawable-canvas
opencv-python
tensorflow
numpy
Pillow
